Raise custom `NotModified` exception directly in `test_exception_header_forwarded`'s `Foo1.get` rather than going through `flask_abort`
The `flask_abort` name is an alias for `flask.abort` which is really `werkzeug.exceptions.abort`, which changed from an `Aborter` instance to a funciton around `Werkzeug==0.11.12`. This is causing the `test_exception_header_forwarded` unittest to fail because functions don't have a `.mapping` attribute...

Chose to remove the `flask.abort` middleman since it wasn't adding anything to the test: it was adding an extra layer of indirection for getting the exception to raise. Test seems to be ensure etags (or other custom header) get set on exceptions, not that those exceptions can be raised with `flask.abort`